Move popup to component
