+++ client/my-sites/plans-grid/components/plan-type-selector/components/popup-messages.tsx
@@ -1,0 +1,133 @@
+import { Popover } from '@automattic/components';
+import styled from '@emotion/styled';
+import * as React from 'react';
+import CSSTransition from 'react-transition-group/CSSTransition';
+
+const StyledPopover = styled( Popover )`
+	&.popover {
+		display: none;
+		opacity: 0;
+		transition-property: opacity, transform;
+		transition-timing-function: ease-in;
+
+		&.popover-enter-active {
+			opacity: 1;
+			transition-duration: 0.3s;
+		}
+
+		&.popover-exit,
+		&.popover-enter-done {
+			opacity: 1;
+			transition-duration: 0.01s;
+		}
+
+		&.is-right,
+		.rtl &.is-left {
+			@media ( min-width: 960px ) {
+				display: block;
+			}
+
+			&.popover-enter {
+				transform: translate( 30px, 0 );
+			}
+
+			&.popover-enter-active,
+			&.popover-enter-done {
+				transform: translate( 20px, 0 );
+			}
+
+			.popover__arrow {
+				border-right-color: var( --color-neutral-100 );
+				&::before {
+					border-right-color: var( --color-neutral-100 );
+				}
+			}
+		}
+
+		.rtl &.is-left {
+			.popover__arrow {
+				right: 40px;
+				border-left-color: var( --color-neutral-100 );
+				&::before {
+					border-left-color: var( --color-neutral-100 );
+				}
+			}
+
+			.popover__inner {
+				left: -50px;
+			}
+		}
+
+		&.is-bottom {
+			@media ( max-width: 960px ) {
+				display: block;
+			}
+
+			&.popover-enter {
+				transform: translate( 0, 22px );
+			}
+
+			&.popover-enter-active,
+			&.popover-enter-done {
+				transform: translate( 0, 12px );
+			}
+
+			.popover__arrow {
+				border-bottom-color: var( --color-neutral-100 );
+				&::before {
+					border-bottom-color: var( --color-neutral-100 );
+				}
+			}
+		}
+
+		.rtl &.is-bottom {
+			.popover__arrow {
+				border-right-color: transparent;
+			}
+		}
+
+		.popover__inner {
+			padding: 8px 10px;
+			max-width: 210px;
+			background-color: var( --color-neutral-100 );
+			border-color: var( --color-neutral-100 );
+			color: var( --color-neutral-0 );
+			border-radius: 2px;
+			text-align: left;
+		}
+	}
+`;
+
+type PopupMessageProps = {
+	context?: HTMLElement;
+	isVisible: boolean;
+	children?: React.ReactNode;
+};
+
+const PopupMessages: React.FunctionComponent< PopupMessageProps > = ( {
+	context,
+	children,
+	isVisible,
+} ) => {
+	const timeout = { enter: 300, exit: 10 };
+	const inProp = Boolean( isVisible && context );
+
+	return (
+		<>
+			{ [ 'right', 'bottom' ].map( ( pos ) => (
+				<CSSTransition key={ pos } in={ inProp } timeout={ timeout } classNames="popover">
+					<StyledPopover
+						position={ pos }
+						context={ context }
+						isVisible={ true }
+						autoPosition={ false }
+					>
+						{ children }
+					</StyledPopover>
+				</CSSTransition>
+			) ) }
+		</>
+	);
+};
+
+export default PopupMessages;
